# Bench `base58` encoding and remove `SmallVec` to improve perf