--- a/base58/src/lib.rs
+++ b/base58/src/lib.rs
@@ -178,9 +178,7 @@ where
     }
 
     // ... then reverse it and convert to chars
-    for _ in 0..leading_zero_count {
-        ret.push(0);
-    }
+    ret.resize(ret.len() + leading_zero_count, 0);
 
     for ch in ret.iter().rev() {
         writer.write_char(BASE58_CHARS[*ch as usize] as char)?;