@apoelstra 2024-05-10
utACK deeb160b86e5e025ea98494b5d30611e47dc8769 -- I also see a 10% improvement and the `base58` crate is already inundated with `Vec`s and requires an allocator. (Maybe we were hoping to move to a smallvec-based scheme eventually and drop the allocator, but there's lots of work to do there and base58 is not where I'd start..)

However, amusingly you have unearthed a clippy warning:
```
error: it looks like the same item is being pushed into this Vec
   --> base58/src/lib.rs:182:9
    |
182 |         ret.push(0);
    |         ^^^
    |
    = help: consider using vec![0;SIZE] or ret.resize(NEW_SIZE, 0)
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#same_item_push
    = note: `-D clippy::same-item-push` implied by `-D warnings`
    = help: to override `-D warnings` add `#[allow(clippy::same_item_push)]`
 
error: could not compile `base58ck` (lib) due to 1 previous error
```

This is not in the code you directly touched, but it did not trigger before because it is `Vec`-specific and we used to use `SmallVec`.